requirements.txt
psycopg2-binary==2.9.9
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
